Implementation update
-resolves all implementation comments
