+++ q2_quality_control/quality_control.py
@@ -128,114 +128,3 @@ def evaluate_taxonomy(output_dir: str, expected_taxa: pd.DataFrame,
                misclassifications=None, underclassifications=None,
                composition_regression=None, score_plot=score_plot,
                mismatch_histogram=None, alignments=None)
-
-
-# decontam added code
-_WHOLE_NUM = (lambda x: x >= 0, 'non-negative')
-_PER_NUM = (lambda x: 1 >= x >= 0, 'between 0 and 1')
-_COL_STR = (lambda x: x in {'column_name', 'column_number'},
-            'sample_name or column_name or column_number')
-_DECON_METHOD_STR = (lambda x: x in {'frequency', 'prevalence', 'combined'},
-                     'freqeuncy, prevalence, combined')
-_BOOLEAN = (lambda x: type(x) is bool, 'True or False')
-# Better to choose to skip, than to implicitly ignore things that KeyError
-_SKIP = (lambda x: True, '')
-_valid_inputs = {
-    'table': _SKIP,
-    'metadata': _SKIP,
-    'threshold': _PER_NUM,
-    'method': _DECON_METHOD_STR,
-    'freq_concentration_column': _SKIP,
-    'prev_control_column': _SKIP,
-    'prev_control_indicator': _SKIP,
-}
-
-
-def _check_inputs(**kwargs):
-    for param, arg in kwargs.items():
-        check_is_valid, explanation = _valid_inputs[param]
-        if not check_is_valid(arg):
-            raise ValueError('Argument to %r was %r, should be %s.'
-                             % (param, arg, explanation))
-
-
-def _decontam_identify_helper(track_fp, method):
-    df = pd.read_csv(track_fp, sep='\t', index_col=0)
-    df.index.name = '#OTU ID'
-    # removes last column containing true/false information from the dataframe
-    df = df.drop(df.columns[[(len(df.columns)-1)]], axis=1)
-    if method == 'combined':
-        df = df.fillna(0)
-
-    # removes all columns that are completely empty
-    temp_transposed_table = df.transpose()
-    temp_transposed_table = temp_transposed_table.dropna()
-    df = temp_transposed_table.transpose()
-    metadata = transform(df, from_type=pd.DataFrame,
-                         to_type=DecontamScoreFormat)
-
-    return metadata
-
-
-def decontam_identify(table: pd.DataFrame,
-                      metadata: qiime2.Metadata,
-                      method: str = 'prevalence',
-                      freq_concentration_column: str = 'NULL',
-                      prev_control_column: str = 'NULL',
-                      prev_control_indicator: str = 'NULL'
-                      ) -> (DecontamScoreFormat):
-    _check_inputs(**locals())
-    with tempfile.TemporaryDirectory() as temp_dir_name:
-        track_fp = os.path.join(temp_dir_name, 'track.tsv')
-        ASV_dest = os.path.join(temp_dir_name, 'temp_ASV_table.csv')
-        table.to_csv(os.path.join(ASV_dest))
-        metadata = metadata.to_dataframe()
-        meta_dest = os.path.join(temp_dir_name, 'temp_metadata.csv')
-        metadata.to_csv(os.path.join(meta_dest))
-
-        cmd = ['run_decontam.R',
-               '--asv_table_path', str(ASV_dest),
-               '--threshold', str(0.1),
-               '--decon_method', method,
-               '--output_track', track_fp,
-               '--meta_table_path', str(meta_dest),
-               '--freq_con_column', str(freq_concentration_column),
-               '--prev_control_or_exp_sample_column',
-               str(prev_control_column),
-               '--prev_control_sample_indicator',
-               str(prev_control_indicator)]
-        try:
-            _run_command(cmd)
-        except subprocess.CalledProcessError as e:
-            if e.returncode == 2:
-                raise ValueError("There was an issue running "
-                                 "run_decontam.R please check your inputs")
-            else:
-                raise Exception("An error was encountered "
-                                "while running Decontam in R "
-                                "(return code %d), please inspect stdout"
-                                " and stderr to learn more." % e.returncode)
-        return _decontam_identify_helper(track_fp, method)
-
-
-def decontam_remove(decontam_scores: qiime2.Metadata,
-                    table: pd.DataFrame,
-                    threshold: float = 0.1
-                    ) -> (biom.Table):
-    with tempfile.TemporaryDirectory() as temp_dir_name:
-        df = decontam_scores.to_dataframe()
-        df.loc[(df['p'].astype(float) <= threshold),
-               'contaminant_seq'] = 'True'
-        df.loc[(df['p'].astype(float) > threshold),
-               'contaminant_seq'] = 'False'
-        df = df[df.contaminant_seq == 'True']
-        remove_these = df.index
-        trans_table = table.transpose()
-        for bad_seq in list(remove_these):
-            trans_table = (trans_table[trans_table.index != bad_seq])
-        output = os.path.join(temp_dir_name, 'temp.tsv.biom')
-        temp_transposed_table = trans_table.transpose()
-        temp_transposed_table.to_csv(output, sep="\t")
-        with open(output) as fh:
-            no_contam_table = biom.Table.from_tsv(fh, None, None, None)
-        return no_contam_table
